+++ lib/fireauth.dart
@@ -8,8 +8,10 @@ import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:provider/provider.dart';
-
 import 'oauth.dart';
+
+//Exports
+export 'auth_controller.dart';
 
 /// Initializes Firebase for Mobile Devices
 ///
@@ -87,9 +89,6 @@ class FireAuthProvider extends ChangeNotifier {
 
       //======================HOT RESTART BUG BYPASS===========================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(
-        recievedUser,
-      );
       //======================HOT RESTART BUG BYPASS===========================
 
       if (onSignInSuccessful != null && recievedUser != null)
@@ -121,9 +120,6 @@ class FireAuthProvider extends ChangeNotifier {
 
       //======================HOT RESTART BUG BYPASS============================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(
-        userCred.user,
-      );
       //======================HOT RESTART BUG BYPASS============================
 
       if (onSignInSuccessful != null && userCred != null)
@@ -153,9 +149,6 @@ class FireAuthProvider extends ChangeNotifier {
 
       //======================HOT RESTART BUG BYPASS============================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(
-        userCred.user,
-      );
       //======================HOT RESTART BUG BYPASS============================
 
       if (onRegisterSuccessful != null && userCred != null)
@@ -184,9 +177,6 @@ class FireAuthProvider extends ChangeNotifier {
 
       //======================HOT RESTART BUG BYPASS============================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(
-        userCred.user,
-      );
       //======================HOT RESTART BUG BYPASS============================
 
       if (onSignInSuccessful != null && userCred != null)
@@ -226,7 +216,7 @@ class FireAuthProvider extends ChangeNotifier {
       bool closeVerificationPopupAfterSubmit = true,
     }) {
       return AlertDialog(
-        title: Text('Verify Phone'),
+        title: Text('OTP Verification'),
         content: Container(
           child: TextField(
             controller: ctr,
@@ -245,9 +235,6 @@ class FireAuthProvider extends ChangeNotifier {
                 if (isDone) {
                   //======================HOT RESTART BUG BYPASS===============
                   await HotRestartBypassMechanism.saveLoginState(true);
-                  await HotRestartBypassMechanism.saveUserInformation(
-                    userCred.user,
-                  );
                   //======================HOT RESTART BUG BYPASS===============
                   if (!closeVerificationPopupAfterSubmit)
                     Navigator.pop(context);
@@ -431,7 +418,6 @@ class FireAuthProvider extends ChangeNotifier {
       if (onSignInSuccessful != null) onSignInSuccessful(user);
       //======================HOT RESTART BUG BYPASS============================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(user);
       //======================HOT RESTART BUG BYPASS============================
     }
     return user;
@@ -452,10 +438,28 @@ class FireAuthProvider extends ChangeNotifier {
       if (onSignInSuccessful != null) onSignInSuccessful(user);
       //======================HOT RESTART BUG BYPASS============================
       await HotRestartBypassMechanism.saveLoginState(true);
-      await HotRestartBypassMechanism.saveUserInformation(user);
       //======================HOT RESTART BUG BYPASS============================
     }
+    return user;
+  }
 
+  //Microsoft
+  Future<User> signInWithMicrosoft({
+    Function(String) onError,
+    Function(User) onSignInSuccessful,
+    bool enableWaitingScreen,
+  }) async {
+    if (enableWaitingScreen) isWaitingForSignInCompletion = true;
+    User user = await OAuthEngine.microsoftOAuthLogin(
+      onError: onError,
+    );
+    isWaitingForSignInCompletion = false;
+    if (user != null) {
+      if (onSignInSuccessful != null) onSignInSuccessful(user);
+      //======================HOT RESTART BUG BYPASS============================
+      await HotRestartBypassMechanism.saveLoginState(true);
+      //======================HOT RESTART BUG BYPASS============================
+    }
     return user;
   }
 
@@ -467,300 +471,6 @@ class FireAuthProvider extends ChangeNotifier {
     isWaitingForSignInCompletion = false; //To Update ChangeNotifier
     await HotRestartBypassMechanism.saveLoginState(false);
     print("Logged Out");
-  }
-}
-
-///This class contains static methods that help you access the Authentication methods effectively
-class AuthController {
-  ///Initates a GoogleSignIn and the AuthManager changes the screen to the destinationFragment
-  ///
-  ///Shows a traditional OAuth Pop up window on the Web and a normal GoogleSignIn Modal on android
-  ///
-  ///Please ensure you have added your SHA-1 Key to Firebase to work on Mobile and ensure to have your
-  ///GoogleSignInClientID in a meta tag in /web/index.html.
-  ///
-  ///Enable Google Authentication in your Firebase Authentication Console for this to work
-  ///
-  ///[context] is neccessary
-  ///
-  ///[signInWithRedirect] (default false) is a boolean that is Flutter Web only and basically allows you to chose if you want your
-  ///OAuth Screen to be a popup or a redirect. Setting this to true, will use a redirect
-  ///
-  ///[enableWaitingScreen] (default false) is a boolean that enables or disables the AuthManager's Waiting Screen
-  ///Until the signIn is complete, the AuthManager will show a default waitingScreen or a custom WaitingScreen depending
-  ///on how you have setup your AuthManager.
-  ///
-  ///[onError] a Callback for any Error that may occur
-  ///
-  ///[onSignInSuccessful] a Callback to perform any action after a successful SignIn
-  static Future<User> signInWithGoogle(
-    BuildContext context, {
-    bool signInWithRedirect,
-    bool enableWaitingScreen,
-    Function(String) onError,
-    Function onSignInSuccessful,
-  }) async {
-    return await Provider.of<FireAuthProvider>(
-      context,
-      listen: false,
-    ).signInWithGoogle(
-      allowSignInWithRedirect: signInWithRedirect ?? false,
-      enableWaitingScreen: enableWaitingScreen ?? false,
-      onError: onError,
-      onSignInSuccessful: onSignInSuccessful,
-    );
-  }
-
-  ///Initiates and Anonymous SignIn and the AuthManager changes the screen to the destinationFragment
-  ///
-  ///Enable Anonymous Authentication in your Firebase Authentication Console for this to work
-  ///[context] is neccessary
-  ///
-  ///[enableWaitingScreen] (default false) is a boolean that enables or disables the AuthManager's Waiting Screen
-  ///Until the signIn is complete, the AuthManager will show a default waitingScreen or a custom WaitingScreen depending
-  ///on how you have setup your AuthManager.
-  ///
-  ///[onSignInSuccessful] a Callback to perform any action after a successful SignIn
-  static Future<User> signInAnonymously(
-    BuildContext context, {
-    bool enableWaitingScreen,
-    Function(User) onSignInSuccessful,
-    Function(String) onError,
-  }) async {
-    return await Provider.of<FireAuthProvider>(
-      context,
-      listen: false,
-    ).signInAnonymously(
-      enableWaitingScreen: enableWaitingScreen ?? true,
-      onSignInSuccessful: onSignInSuccessful,
-      onError: onError,
-    );
-  }
-
-  ///Registers the Email and Password Combination on Firebase
-  ///initiates a login and the AuthManager changes the screen to the destinationFragment
-  ///Call this to create a new User and instantly log them in
-  ///
-  ///Please enable Email and Password Authentication in your Firebase Authentication Console for this to work
-  ///
-  ///[context] is necessary
-  ///
-  ///[email] is required and is self-explanatory
-  ///
-  ///[password] is also required and is self-explanatory
-  ///
-  ///[onError] is a callback to handle any errors during the process
-  ///
-  ///[onRegisterSuccessful] a Callback to perform any action after a successful SignIn
-  static Future<User> registerWithEmailAndPassword(
-    BuildContext context, {
-    @required String email,
-    @required String password,
-    Function(String) onError,
-    Function(User) onRegisterSuccessful,
-  }) async {
-    return await Provider.of<FireAuthProvider>(context, listen: false)
-        .registerWithEmailAndPassword(
-      email: email,
-      password: password,
-      onError: onError,
-      onRegisterSuccessful: onRegisterSuccessful,
-    );
-  }
-
-  ///Initates a Mail & Password SignIn and the AuthManager changes the screen to the destinationFragment
-  ///
-  ///Please enable Email and Password Authentication in your Firebase Authentication Console for this to work
-  ///
-  ///[context] is necessary
-  ///
-  ///[email] is required and is self-explanatory
-  ///
-  ///[password] is also required and is self-explanatory
-  ///
-  ///[onError] is a callback to handle any errors during the process
-  ///
-  ///[onIncorrectCredentials] is a callback to handle incorrect credentials
-  ///
-  ///[onSignInSuccessful] a Callback to perform any action after a successful SignIn
-  static Future<User> signInWithEmailAndPassword(
-    BuildContext context, {
-    @required String email,
-    @required String password,
-    Function(String) onError,
-    Function onIncorrectCredentials,
-    Function(User) onSignInSuccessful,
-  }) async {
-    return await Provider.of<FireAuthProvider>(context, listen: false)
-        .signInWithEmailAndPassword(
-      email: email,
-      password: password,
-      onError: onError,
-      onIncorrectCredentials: onIncorrectCredentials,
-      onSignInSuccessful: onSignInSuccessful,
-    );
-  }
-
-  ///Initiates a PhoneNumber SignIn, brings up an OTPVerificationModal and then the AuthManager changes the screen to the destinationFragment
-  ///
-  ///Please enable Phone Authentication in your Firebase Authentication Console for this to work
-  ///
-  ///Please ensure you have added your SHA-1 and SHA-256 Keys to Firebase
-  ///
-  ///If you want to avoid ReCaptcha Redirect verification, Go to Google Cloud Console, Open your firebase
-  ///project, search Android Device Verification and enable it
-  ///
-  ///The Phone Number must be in the format: +\<country code>\<phone number>
-  ///
-  ///[context] is necessary
-  ///
-  ///[phoneNumber] is the phoneNumber used to login
-  ///
-  ///[onError] is a callback to handle any errors in this process
-  ///
-  ///[onInvalidVerificationCode] is a callback to handle the situation where the OTP is incorrect
-  ///
-  ///[onSignInSuccessful] a Callback to perform any action after a successful SignIn
-  ///
-  ///[closeVerificationPopupAfterSubmit] a boolean which when true, closes the OTP Verification Dialog after an attempt and if
-  ///false leaves it open, basically if the User Wants to Retry
-  static Future<User> signInWithPhoneNumber(
-    BuildContext context, {
-    String phoneNumber,
-    Function(String) onError,
-    Function onInvalidVerificationCode,
-    Function(User) onSignInSuccessful,
-    bool closeVerificationPopupAfterSubmit = false,
-  }) async {
-    final provider = Provider.of<FireAuthProvider>(context, listen: false);
-
-    return await provider.signInWithPhoneNumber(
-      context: context,
-      phoneNumber: phoneNumber,
-      onError: onError,
-      onInvalidVerificationCode: onInvalidVerificationCode,
-      onSignInSuccessful: onSignInSuccessful,
-      closeVerificationPopupAfterSubmit: closeVerificationPopupAfterSubmit,
-    );
-  }
-
-  ///Initiates a Twitter OAuth SignUp Flow based on the OAuthEngine Implementation
-  ///
-  ///[context] is necessary
-  ///
-  ///[onError] is a callback that is invoked when an error is encountered
-  ///
-  ///[onSignInSuccessful] is a callback that is invoked when the signIn is successful.
-  ///It provides a User which you can use to perform other actions.
-  ///
-  ///[enableWaitingScreen] (default false) is a boolean that enables or disables the AuthManager's Waiting Screen
-  ///Until the signIn is complete, the AuthManager will show a default waitingScreen or a custom WaitingScreen depending
-  ///on how you have setup your AuthManager.
-  ///
-  ///Setup Process
-  ///
-  ///Enable Twitter Authentication in the Firebase Authentication Console
-  ///
-  ///Open the Twitter Developer Console, Create a new app, add its API Key & Secret
-  ///to the respective fields in the TwitterAuthDialog in the FirebaseConsole.
-  ///
-  ///Then Copy the callbackURL from the console, Enable 3-Legged-OAuth in the Twitter
-  ///Developer console and add the callbackURL there and the setup is done!
-  static Future<User> signInWithTwiter(
-    BuildContext context, {
-    Function(String) onError,
-    Function(User) onSignInSuccessful,
-    bool enableWaitingScreen,
-  }) async {
-    final provider = Provider.of<FireAuthProvider>(context, listen: false);
-    return await provider.signInWithTwitter(
-      onError: onError,
-      onSignInSuccessful: onSignInSuccessful,
-      enableWaitingScreen: enableWaitingScreen ?? true,
-    );
-  }
-
-  ///Initiates a Github OAuth SignUp Flow based on the OAuthEngine Implementation
-  ///
-  ///[context] is necessary
-  ///
-  ///[onError] is a callback that is invoked when an error is encountered
-  ///
-  ///[onSignInSuccessful] is a callback that is invoked when the signIn is successful.
-  ///It provides a User which you can use to perform other actions.
-  ///
-  ///[enableWaitingScreen] (default false) is a boolean that enables or disables the AuthManager's Waiting Screen
-  ///Until the signIn is complete, the AuthManager will show a default waitingScreen or a custom WaitingScreen depending
-  ///on how you have setup your AuthManager.
-  ///
-  ///Setup Process
-  ///
-  ///Enable Github Authentication in the Firebase Console, Create a Github OAuth App, copy the client ID
-  ///and client secret and paste it in the dialog, copy the callback URL from the dialog and paste it where needed
-  ///in your Github OAuth App Configuration
-  static Future<User> signInWithGithub(
-    BuildContext context, {
-    Function(String) onError,
-    Function(User) onSignInSuccessful,
-    bool enableWaitingScreen,
-  }) async {
-    final provider = Provider.of<FireAuthProvider>(context, listen: false);
-    return await provider.signInWithGithub(
-      onError: onError,
-      onSignInSuccessful: onSignInSuccessful,
-      enableWaitingScreen: enableWaitingScreen ?? true,
-    );
-  }
-
-  /// Initiates a logout and the authManager redirects to the loginFragment
-  ///
-  /// [context] is necessary
-  ///
-  /// [onLogout] is a callback function that is called immediately after a logout
-  static logout(BuildContext context, {Function onLogout}) {
-    Provider.of<FireAuthProvider>(context, listen: false).logout(
-      onLogout: onLogout,
-    );
-  }
-
-  /// An easy way to get the Currently Logged in User
-  ///
-  /// [context] is necesssary
-  /// [customMapping] is an optional Function accepting a User as its arguement and returns it in
-  /// a format of your choice, for example creating a custom User Model.
-  ///
-  /// The recieved User will have all its data filled for GoogleSignIn only, for other forms of Auth
-  /// Some of the User Data for example email, displayName etc will be empty.
-  ///
-  /// emailVerified is true only for GoogleSignIn
-  ///
-  /// Example:
-  /// ```dart
-  ///  AuthController.getCurrentUser(
-  ///   context,
-  ///   customMapping: (user) => {
-  ///     'name': user.displayName,
-  ///     'email': user.email,
-  ///   },
-  /// );
-  /// ```
-  static User getCurrentUser(BuildContext context,
-      {Function(User) customMapping}) {
-    final provider = Provider.of<FireAuthProvider>(context, listen: false);
-    User cUser = provider.authInstance.currentUser;
-    // Enable this, if currentUser is null during testing even though user is logged in,
-    // if (cUser == null) {
-    //HotRestartBug Correction
-    //   if (Foundation.kDebugMode && Foundation.kIsWeb) {
-    //     print(
-    //       "Returning HotRestartBypassMechanism:SavedUser as User was null. (DartWebSDKBug)",
-    //     );
-    //     return await HotRestartBypassMechanism.getUserInformation();
-    //   }
-    //   return null;
-    // }
-    if (customMapping != null) return customMapping(cUser);
-    return cUser;
   }
 }